+++ Changes.md
@@ -4,6 +4,7 @@ Change Log in version 3.11.0.5(2021081004)
    Thus for guests and authenticated users, when 'un-ticked' then the 'side-post' region does not show on the front page, but it
    does for site admins.  When 'ticked' then 'side-post' will show for all and have content when there is some.
 2. Fix 'No middle region'.
+3. Tidied 'get_block_regions()' with the creation of the 'adaptable-block-area' style.
 
 Change Log in version 3.11.0.4 (2021081003)
 ===========================================